Update version of api, added manyToMany implementation, added user profile

+++ em/src/main/java/com/em/service/UserService.java
@@ -10,7 +10,7 @@ public interface UserService {
 	
 	public List<User> getAllUsers();
 	
-	public User getUser(Long id);
+	public User getUser(Integer id);
 	
 	public void addUser(User user);
